integrated mess all options in secretary page

+++ crs/login/models.py
@@ -253,8 +253,7 @@ class Fooditems(models.Model):
     vitamins = models.IntegerField()
     proteins = models.IntegerField()
     fat = models.IntegerField()
-    nutritions = models.DecimalField(max_digits=4, decimal_places=2)
-
+    nutritions = models.FloatField()
     class Meta:
         managed = False
         db_table = 'foodItems'
